+++ internal/user/full_sync.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"errors"
 	"github.com/openimsdk/openim-sdk-core/v3/pkg/db/model_struct"
+	userPb "github.com/openimsdk/protocol/user"
 	"github.com/openimsdk/tools/errs"
 	"github.com/openimsdk/tools/log"
 	"github.com/openimsdk/tools/utils/datautil"
@@ -52,7 +53,7 @@ func (u *User) SyncAllCommandWithoutNotice(ctx context.Context) error {
 }
 
 func (u *User) syncAllCommand(ctx context.Context, withNotice bool) error {
-	resp, err := u.processUserCommandGetAll(ctx)
+	resp, err := u.processUserCommandGetAll(ctx, &userPb.ProcessUserCommandGetAllReq{UserID: u.loginUserID})
 	if err != nil {
 		return err
 	}
@@ -62,8 +63,8 @@ func (u *User) syncAllCommand(ctx context.Context, withNotice bool) error {
 	}
 	log.ZDebug(ctx, "sync command", "data from server", resp, "data from local", localData)
 	if withNotice {
-		return u.commandSyncer.Sync(ctx, datautil.Batch(ServerCommandToLocalCommand, resp), localData, nil)
+		return u.commandSyncer.Sync(ctx, datautil.Batch(ServerCommandToLocalCommand, resp.CommandResp), localData, nil)
 	} else {
-		return u.commandSyncer.Sync(ctx, datautil.Batch(ServerCommandToLocalCommand, resp), localData, nil, false, true)
+		return u.commandSyncer.Sync(ctx, datautil.Batch(ServerCommandToLocalCommand, resp.CommandResp), localData, nil, false, true)
 	}
 }
