+++ internal/user/sdk.go
@@ -13,47 +13,3 @@
 // limitations under the License.
 
 package user
-
-import (
-	"context"
-	"github.com/openimsdk/openim-sdk-core/v3/pkg/constant"
-	"github.com/openimsdk/openim-sdk-core/v3/pkg/db/model_struct"
-	"github.com/openimsdk/protocol/sdkws"
-)
-
-func (u *User) GetUsersInfo(ctx context.Context, userIDs []string) ([]*model_struct.LocalUser, error) {
-	return u.GetUsersInfoFromSvr(ctx, userIDs)
-}
-
-func (u *User) GetSelfUserInfo(ctx context.Context) (*model_struct.LocalUser, error) {
-	return u.getSelfUserInfo(ctx)
-}
-
-// Deprecated: user SetSelfInfoEx instead
-func (u *User) SetSelfInfo(ctx context.Context, userInfo *sdkws.UserInfo) error {
-	return u.updateSelfUserInfo(ctx, userInfo)
-}
-func (u *User) SetSelfInfoEx(ctx context.Context, userInfo *sdkws.UserInfoWithEx) error {
-	return u.updateSelfUserInfoEx(ctx, userInfo)
-}
-func (u *User) SetGlobalRecvMessageOpt(ctx context.Context, opt int) error {
-	if err := u.setGlobalRecvMessageOpt(ctx, int32(opt)); err != nil {
-		return err
-	}
-	u.SyncLoginUserInfo(ctx)
-	return nil
-}
-
-func (u *User) UpdateMsgSenderInfo(ctx context.Context, nickname, faceURL string) (err error) {
-	if nickname != "" {
-		if err = u.DataBase.UpdateMsgSenderNickname(ctx, u.loginUserID, nickname, constant.SingleChatType); err != nil {
-			return err
-		}
-	}
-	if faceURL != "" {
-		if err = u.DataBase.UpdateMsgSenderFaceURL(ctx, u.loginUserID, faceURL, constant.SingleChatType); err != nil {
-			return err
-		}
-	}
-	return nil
-}
