+++ internal/user/user.go
@@ -24,8 +24,6 @@ import (
 	"github.com/openimsdk/openim-sdk-core/v3/pkg/syncer"
 	"github.com/openimsdk/openim-sdk-core/v3/sdk_struct"
 	"github.com/openimsdk/protocol/sdkws"
-	userPb "github.com/openimsdk/protocol/user"
-	"github.com/openimsdk/tools/log"
 	"github.com/openimsdk/tools/utils/datautil"
 
 	"github.com/openimsdk/openim-sdk-core/v3/open_im_sdk_callback"
@@ -33,6 +31,15 @@ import (
 	"github.com/openimsdk/openim-sdk-core/v3/pkg/constant"
 	"github.com/openimsdk/openim-sdk-core/v3/pkg/utils"
 )
+
+// NewUser creates a new User object.
+func NewUser(dataBase db_interface.DataBase, loginUserID string, conversationCh chan common.Cmd2Value) *User {
+	user := &User{DataBase: dataBase, loginUserID: loginUserID, conversationCh: conversationCh}
+	user.initSyncer()
+	user.UserBasicCache = cache.NewCache[string, *sdk_struct.BasicInfo]()
+	//user.OnlineStatusCache = cache.NewCache[string, *userPb.OnlineStatus]()
+	return user
+}
 
 // User is a struct that represents a user in the system.
 type User struct {
@@ -50,30 +57,6 @@ type User struct {
 // SetListener sets the user's listener.
 func (u *User) SetListener(listener func() open_im_sdk_callback.OnUserListener) {
 	u.listener = listener
-}
-
-func (u *User) UserOnlineStatusChange(users map[string][]int32) {
-	for userID, onlinePlatformIDs := range users {
-		status := userPb.OnlineStatus{
-			UserID:      userID,
-			PlatformIDs: onlinePlatformIDs,
-		}
-		if len(status.PlatformIDs) == 0 {
-			status.Status = constant.Offline
-		} else {
-			status.Status = constant.Online
-		}
-		u.listener().OnUserStatusChanged(utils.StructToJsonString(&status))
-	}
-}
-
-// NewUser creates a new User object.
-func NewUser(dataBase db_interface.DataBase, loginUserID string, conversationCh chan common.Cmd2Value) *User {
-	user := &User{DataBase: dataBase, loginUserID: loginUserID, conversationCh: conversationCh}
-	user.initSyncer()
-	user.UserBasicCache = cache.NewCache[string, *sdk_struct.BasicInfo]()
-	//user.OnlineStatusCache = cache.NewCache[string, *userPb.OnlineStatus]()
-	return user
 }
 
 func (u *User) initSyncer() {
@@ -150,76 +133,6 @@ func (u *User) initSyncer() {
 	)
 }
 
-// DoNotification handles incoming notifications for the user.
-func (u *User) DoNotification(ctx context.Context, msg *sdkws.MsgData) {
-	log.ZDebug(ctx, "user notification", "msg", *msg)
-	go func() {
-		switch msg.ContentType {
-		case constant.UserInfoUpdatedNotification:
-			u.userInfoUpdatedNotification(ctx, msg)
-		case constant.UserStatusChangeNotification:
-			//u.userStatusChangeNotification(ctx, msg)
-		case constant.UserCommandAddNotification:
-			u.userCommandAddNotification(ctx, msg)
-		case constant.UserCommandDeleteNotification:
-			u.userCommandDeleteNotification(ctx, msg)
-		case constant.UserCommandUpdateNotification:
-			u.userCommandUpdateNotification(ctx, msg)
-		default:
-			// log.Error(operationID, "type failed ", msg.ClientMsgID, msg.ServerMsgID, msg.ContentType)
-		}
-	}()
-}
-
-// userInfoUpdatedNotification handles notifications about updated user information.
-func (u *User) userInfoUpdatedNotification(ctx context.Context, msg *sdkws.MsgData) {
-	log.ZDebug(ctx, "userInfoUpdatedNotification", "msg", *msg)
-	tips := sdkws.UserInfoUpdatedTips{}
-	if err := utils.UnmarshalNotificationElem(msg.Content, &tips); err != nil {
-		log.ZError(ctx, "comm.UnmarshalTips failed", err, "msg", msg.Content)
-		return
-	}
-
-	if tips.UserID == u.loginUserID {
-		u.SyncLoginUserInfo(ctx)
-	} else {
-		log.ZDebug(ctx, "detail.UserID != u.loginUserID, do nothing", "detail.UserID", tips.UserID, "u.loginUserID", u.loginUserID)
-	}
-}
-
-// userCommandAddNotification handle notification when user add favorite
-func (u *User) userCommandAddNotification(ctx context.Context, msg *sdkws.MsgData) {
-	log.ZDebug(ctx, "userCommandAddNotification", "msg", *msg)
-	tip := sdkws.UserCommandAddTips{}
-	if tip.ToUserID == u.loginUserID {
-		u.SyncAllCommand(ctx)
-	} else {
-		log.ZDebug(ctx, "ToUserID != u.loginUserID, do nothing", "detail.UserID", tip.ToUserID, "u.loginUserID", u.loginUserID)
-	}
-}
-
-// userCommandDeleteNotification handle notification when user delete favorite
-func (u *User) userCommandDeleteNotification(ctx context.Context, msg *sdkws.MsgData) {
-	log.ZDebug(ctx, "userCommandAddNotification", "msg", *msg)
-	tip := sdkws.UserCommandDeleteTips{}
-	if tip.ToUserID == u.loginUserID {
-		u.SyncAllCommand(ctx)
-	} else {
-		log.ZDebug(ctx, "ToUserID != u.loginUserID, do nothing", "detail.UserID", tip.ToUserID, "u.loginUserID", u.loginUserID)
-	}
-}
-
-// userCommandUpdateNotification handle notification when user update favorite
-func (u *User) userCommandUpdateNotification(ctx context.Context, msg *sdkws.MsgData) {
-	log.ZDebug(ctx, "userCommandAddNotification", "msg", *msg)
-	tip := sdkws.UserCommandUpdateTips{}
-	if tip.ToUserID == u.loginUserID {
-		u.SyncAllCommand(ctx)
-	} else {
-		log.ZDebug(ctx, "ToUserID != u.loginUserID, do nothing", "detail.UserID", tip.ToUserID, "u.loginUserID", u.loginUserID)
-	}
-}
-
 // GetUsersInfoFromSvr retrieves user information from the server.
 func (u *User) GetUsersInfoFromSvr(ctx context.Context, userIDs []string) ([]*model_struct.LocalUser, error) {
 	users, err := u.getUsersInfo(ctx, userIDs)
@@ -227,18 +140,6 @@ func (u *User) GetUsersInfoFromSvr(ctx context.Context, userIDs []string) ([]*mo
 		return nil, sdkerrs.WrapMsg(err, "GetUsersInfoFromSvr failed")
 	}
 	return datautil.Batch(ServerUserToLocalUser, users), nil
-}
-
-// GetSingleUserFromSvr retrieves user information from the server.
-func (u *User) GetSingleUserFromSvr(ctx context.Context, userID string) (*model_struct.LocalUser, error) {
-	users, err := u.GetUsersInfoFromSvr(ctx, []string{userID})
-	if err != nil {
-		return nil, err
-	}
-	if len(users) > 0 {
-		return users[0], nil
-	}
-	return nil, sdkerrs.ErrUserIDNotFound.WrapMsg(fmt.Sprintf("getSelfUserInfo failed, userID: %s not exist", userID))
 }
 
 // getSelfUserInfo retrieves the user's information.
@@ -274,63 +175,4 @@ func (u *User) updateSelfUserInfoEx(ctx context.Context, userInfo *sdkws.UserInf
 	}
 	_ = u.SyncLoginUserInfo(ctx)
 	return nil
-}
-
-// ProcessUserCommandAdd CRUD user command
-func (u *User) ProcessUserCommandAdd(ctx context.Context, userCommand *userPb.ProcessUserCommandAddReq) error {
-	req := &userPb.ProcessUserCommandAddReq{UserID: u.loginUserID, Type: userCommand.Type, Uuid: userCommand.Uuid, Value: userCommand.Value}
-	if err := u.processUserCommandAdd(ctx, req); err != nil {
-		return err
-	}
-	return u.SyncAllCommand(ctx)
-}
-
-// ProcessUserCommandDelete delete user's choice
-func (u *User) ProcessUserCommandDelete(ctx context.Context, userCommand *userPb.ProcessUserCommandDeleteReq) error {
-	req := &userPb.ProcessUserCommandDeleteReq{UserID: u.loginUserID, Type: userCommand.Type, Uuid: userCommand.Uuid}
-	if err := u.processUserCommandDelete(ctx, req); err != nil {
-		return err
-	}
-	return u.SyncAllCommand(ctx)
-}
-
-// ProcessUserCommandUpdate update user's choice
-func (u *User) ProcessUserCommandUpdate(ctx context.Context, userCommand *userPb.ProcessUserCommandUpdateReq) error {
-	req := &userPb.ProcessUserCommandUpdateReq{UserID: u.loginUserID, Type: userCommand.Type, Uuid: userCommand.Uuid, Value: userCommand.Value}
-	if err := u.processUserCommandUpdate(ctx, req); err != nil {
-		return err
-	}
-	return u.SyncAllCommand(ctx)
-}
-
-// ProcessUserCommandGetAll get user's choice
-func (u *User) ProcessUserCommandGetAll(ctx context.Context) ([]*userPb.CommandInfoResp, error) {
-	localCommands, err := u.DataBase.ProcessUserCommandGetAll(ctx)
-	if err != nil {
-		return nil, err // Handle the error appropriately
-	}
-	var result []*userPb.CommandInfoResp
-	for _, localCommand := range localCommands {
-		result = append(result, &userPb.CommandInfoResp{
-			Type:       localCommand.Type,
-			CreateTime: localCommand.CreateTime,
-			Uuid:       localCommand.Uuid,
-			Value:      localCommand.Value,
-		})
-	}
-	return result, nil
-}
-
-// ParseTokenFromSvr parses a token from the server.
-func (u *User) ParseTokenFromSvr(ctx context.Context) (int64, error) {
-	resp, err := u.parseToken(ctx)
-	if err != nil {
-		return 0, err
-	}
-	return resp.ExpireTimeSeconds, err
-}
-
-// GetServerUserInfo retrieves user information from the server.
-func (u *User) GetServerUserInfo(ctx context.Context, userIDs []string) ([]*sdkws.UserInfo, error) {
-	return u.getUsersInfo(ctx, userIDs)
 }
